misc/HtmlTestSuite/results/test-0231.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta charset="utf-8"/>
<title>Content Visibility: content-visibility inheritance</title>
<link href="mailto:vmpstr@chromium.org" rel="author" title="Vladimir Levin"/>
<link href="https://drafts.csswg.org/css-contain/#content-visibility" rel="help"/>
<meta content="content-visibility does not inherit" name="assert"/>
<meta content="content-visibility has initial value 'visible'" name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/css/support/parsing-testcommon.js"/>
<script src="/css/support/inheritance-testcommon.js"/>
</head>
<body>
<div id="container">
<div id="target"/>
</div>
<script>
assert_not_inherited("content-visibility", "visible", "auto");
</script>
</body>
</html>
